// app/dashboard/profile/edit/types.ts
// Types for Profile Edit Page

export interface FormData {
  email: string;
  password: string;
  passwordConfirm: string;
  name: string;
  phone: string;
  businessType: string;
  companyName: string;
  location: string;
  industry: string;
  employees: string;
  representativeName: string;
  contactName: string;
}